ADTS-278 Fix files subMenu feature
